Cleaned shader.cpp, pl.cpp

+++ src/common/containers.h
@@ -237,9 +237,9 @@ private:
 const	int		TRIE_NODE_LEAF				=	1;	
 
 // Some misc macros
-#define	isLeaf(__node)		(((unsigned long long) __node) & TRIE_NODE_LEAF)
-#define	getLeaf(__node)		(CTrieLeaf *) (((unsigned long long) __node) & (~TRIE_NODE_LEAF))
-#define	makeNode(__node)	(CTrieNode *) (((unsigned long long) __node) | TRIE_NODE_LEAF)
+#define	isLeaf(__node)		(((uintptr_t) __node) & TRIE_NODE_LEAF)
+#define	getLeaf(__node)		(CTrieLeaf *) (((uintptr_t) __node) & (~TRIE_NODE_LEAF))
+#define	makeNode(__node)	(CTrieNode *) (((uintptr_t) __node) | TRIE_NODE_LEAF)
 
 ///////////////////////////////////////////////////////////////////////
 // Class				:	CTrie
